requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
opencv-python>=4.9.0
openai>=1.12.0
pyyaml>=6.0.2
aiosqlite>=0.20.0
python-dotenv>=1.0.1
Pillow>=10.3.0
